requirements.txt
paddlepaddle-gpu==2.4.2.post112
paddlenlp==2.5.2
nltk==3.8.1